+++ src/features/alloc_ext.rs
@@ -9,9 +9,8 @@ use {
 /// These helpers simplify common, not entirely trivial allocation patterns.
 pub trait AllocExt: Alloc {
     #[cfg(not(feature = "clone_to_uninit"))]
+    // TODO: fix outdated docs like this referred to walloc
     /// Allocates memory for a single `T` and clones `data` into it.
-    ///
-    /// This is equivalent to `alloc.`[`walloc`](Self::walloc)`(data.clone()))`
     ///
     /// # Errors
     ///
@@ -22,8 +21,7 @@ pub trait AllocExt: Alloc {
     fn alloc_clone<T: Clone>(&self, data: &T) -> Result<NonNull<T>, AllocError> {
         // clone the data before allocating to avoid needing a guard
         let dup = data.clone();
-        let mem =
-            tri!(do self.alloc(<T as crate::data::type_props::SizedProps>::LAYOUT)).cast();
+        let mem = tri!(do self.alloc(<T as crate::data::type_props::SizedProps>::LAYOUT)).cast();
         // SAFETY: the pointer will have at least enough space for a `T`
         unsafe {
             ptr::write(mem.as_ptr(), dup);
@@ -50,19 +48,14 @@ pub trait AllocExt: Alloc {
         //  the pointer was just allocated by this allocator, is valid for writes of at
         //  least `data`'s size, and aligned.
         unsafe {
-            alloc_then(
-                self,
-                crate::data::type_props::PtrProps::layout(&data),
-                data,
-                |p, data| {
-                    let guard = crate::helpers::AllocGuard::new(
-                        NonNull::new_unchecked(crate::unstable_util::with_meta(p.as_ptr(), data)),
-                        self
-                    );
-                    data.clone_to_uninit(guard.as_ptr().cast::<u8>());
-                    guard.release()
-                }
-            )
+            alloc_then(self, crate::data::type_props::PtrProps::layout(&data), data, |p, data| {
+                let guard = crate::helpers::AllocGuard::new(
+                    NonNull::new_unchecked(crate::unstable_util::with_meta(p.as_ptr(), data)),
+                    self
+                );
+                data.clone_to_uninit(guard.as_ptr().cast::<u8>());
+                guard.release()
+            })
         }
     }
 
@@ -75,9 +68,7 @@ pub trait AllocExt: Alloc {
     /// - [`AllocError::AllocFailed`] if allocation fails.
     /// - [`AllocError::ZeroSizedLayout`] if `data.sz() == 0`.
     #[track_caller]
-    fn alloc_clone<
-        T: core::clone::CloneToUninit + crate::data::type_props::VarSized + ?Sized
-    >(
+    fn alloc_clone<T: core::clone::CloneToUninit + crate::data::type_props::VarSized + ?Sized>(
         &self,
         data: &T
     ) -> Result<NonNull<T>, AllocError> {